Fix static link compilation (#8)
* Use C flags instead of CXX

* Fix warnings from MSVC

* Use target to set compilation flags

* Fix warnings

* Update CMakeLists to ignore warnings in libyara

* Fix last warnings

+++ libyarattd/src/libyarattd_utils.c
@@ -113,7 +113,7 @@ int check_idx_file(const wchar_t* filename)
 {
   // Assumes that the filename is the one of the trace file
   // TODO use magic bytes instead of extension
-  unsigned int len = wcslen(filename);
+  size_t len = wcslen(filename);
   if (len < 3)
     return ERROR_COULD_NOT_OPEN_FILE;
